[IS-20] npm fix
