+++ src/components/OverviewChart/OverviewChart.tsx
@@ -1,0 +1,5 @@
+const OverviewChart = () => {
+  return <div>OverviewChart</div>;
+};
+
+export default OverviewChart;
