+++ src/components/OverviewChart/styled.ts
@@ -1,0 +1,1 @@
+import styled from "styled-components";
